move hasPermission() twig function to permissions module
